Fixed attacking
Squad now checks if a unit is trying to attack another unit via the
getOrder method.. Idk why this works but it does...
Army Manager is now simplified to work with the new Squad attack move.
Strategy Manager is simplified to allow for these changes.

Military manager just commented out a print statement.