These Changes Won our First Game Against Built in!
Now we scout at the very start with an SCV. Changes were made to MM and
StrateManager for this to happen.
There were changes in strategy to allow for better timing on buildings -
    We weren't buliding our supply depot early enough before 10 supply
Changed The debugging tool (showing where things are going to) to not have
lines to (0,0). less clutter this way.