Transitioned Production manager to use the TechPaths
Prod Manager now uses the findTechPath() method to get the path to a unit
type, and then examines the path.

ProcessQueue also will now pop things off the queue if it is not the last
thing in the queue.

This needs to be tested a little more but otherwise seems to be working.