Implimeneted TD-Learning choosing best state
We are now fully wired up with TD-Learning.
We are using the Memory and the state transition to choose which of the
legal Goals we should pick.
So far it seems to be running great, choosing random goals as it goes.
Debugging info is shown on the screen so we know what it is doing.
I had to make some small changes to State transition and it's unit Test
because of the way that the ToString method works on UnitTYpes.