TD-Learning Upgraded to BWAPI 4.1.2
This upgrade was in in response to an issue with the BWTA that caused a bug
and made learning more difficult as well as playing concecutive games.

The Following is a list of all the things needed in order to run this branch:

Install instructions are on the SSCAIT tutorial page. (Also the new version of BWAPI has an installer)

These are the refactors that were done to upgrade to 4.1.2:
- no more c_Str() method. (had to replace toString)
- a few methods had parameter swaps. like 3. had to swap the order of the
parameters for the call. (That's all).
- One Method had to be replaced with it's new equivalence. (it's in a part
of code that isn't used though)
- Had to re put the underscore character back in the state transition class
for the valid transitions.