Revert "TD-Learning Upgraded to BWAPI 4.1.2"
This reverts commit 56a4700b491ba8a8d623bc3160e7fddeaeaf5179.
This reverts the Upgrade to BWAPI 4.1.2 Since that was supposed
to be done in a different branch. This will take us back to the
point in TD-learning where everything was working except for the
region finding from BWTA. BWTA had a bug in it which is what
prompted the upgrade.
Again the Upgrade can be found in the TD_LEARNING_UPGRADE branch.