Can Now Build Machine Shop
Since the Machine Shop is an addon to a building, it is built differently.
This requires the Production manager to build it differently. This coded
did not exist prior to this commit.
Production manager can now build a machine shop if possible.

This enables the TD_Learning to actually build its entire gambit of units.

This will require that training start again from zero games, to re train...